Refactor web, relocate files to web folder
