+++ lib/index.js
@@ -6,7 +6,7 @@ const compression = require('compression');
 const Auth = require('./auth');
 const Logger = require('./logger');
 const Config = require('./config');
-const Middleware = require('./middleware');
+const Middleware = require('./web/middleware');
 const Cats = require('./status-cats');
 const Storage = require('./storage');
 
@@ -77,14 +77,14 @@ module.exports = function(config_hash) {
     });
   }
 
-  app.use(require('./index-api')(config, auth, storage));
+  app.use(require('./web/api/api')(config, auth, storage));
 
   if (config.web && config.web.enable === false) {
     app.get('/', function(req, res, next) {
       next( Error[404]('web interface is disabled in the config file') );
     });
   } else {
-    app.use(require('./index-web')(config, auth, storage));
+    app.use(require('./web/index')(config, auth, storage));
   }
 
   app.get('/*', function(req, res, next) {
