+++ lib/web/api/api.js
@@ -5,9 +5,9 @@ let express = require('express');
 let bodyParser = require('body-parser');
 let Error = require('http-errors');
 let Path = require('path');
-let Middleware = require('./middleware');
-let Notify = require('./notify');
-let Utils = require('./utils');
+let Middleware = require('../middleware');
+let Notify = require('../../notify');
+let Utils = require('../../utils');
 let expect_json = Middleware.expect_json;
 let match = Middleware.match;
 let media = Middleware.media;
