+++ lib/web/index.js
@@ -7,9 +7,9 @@ let express = require('express');
 let fs = require('fs');
 let Handlebars = require('handlebars');
 let renderReadme = require('render-readme');
-let Search = require('./search');
+let Search = require('../search');
 let Middleware = require('./middleware');
-let Utils = require('./utils');
+let Utils = require('../utils');
 let match = Middleware.match;
 let validate_name = Middleware.validate_name;
 let validate_pkg = Middleware.validate_package;
@@ -37,12 +37,12 @@ module.exports = function(config, auth, storage) {
 
   Search.configureStorage(storage);
 
-  Handlebars.registerPartial('entry', fs.readFileSync(require.resolve('./GUI/entry.hbs'), 'utf8'));
+  Handlebars.registerPartial('entry', fs.readFileSync(require.resolve('./ui/entry.hbs'), 'utf8'));
   let template;
   if (config.web && config.web.template) {
     template = Handlebars.compile(fs.readFileSync(config.web.template, 'utf8'));
   } else {
-    template = Handlebars.compile(fs.readFileSync(require.resolve('./GUI/index.hbs'), 'utf8'));
+    template = Handlebars.compile(fs.readFileSync(require.resolve('./ui/index.hbs'), 'utf8'));
   }
   app.get('/', function(req, res, next) {
     let base = Utils.combineBaseUrl(Utils.getWebProtocol(req), req.get('host'), config.url_prefix);
