+++ lib/web/middleware.js
@@ -4,8 +4,8 @@
 
 const crypto = require('crypto');
 const Error = require('http-errors');
-const utils = require('./utils');
-const Logger = require('./logger');
+const utils = require('../utils');
+const Logger = require('../logger');
 
 module.exports.match = function match(regexp) {
   return function(req, res, next, value, name) {
