+++ lib/web/static/favicon.ico
