+++ lib/web/static/fontello.eot
