+++ lib/web/static/fontello.ttf
