+++ lib/web/static/logo-sm.png
