+++ lib/web/static/main.css
@@ -7166,8 +7166,8 @@ Original style from softwaremaniacs.org (c) Ivan Sagalaev <Maniac@SoftwareManiac
 }
 @font-face {
   font-family: 'fontello';
-  src: url('../static/fontello.eot?10872183');
-  src: url('../static/fontello.eot?10872183#iefix') format('embedded-opentype'), url('../static/fontello.woff?10872183') format('woff'), url('../static/fontello.ttf?10872183') format('truetype'), url('../static/fontello.svg?10872183#fontello') format('svg');
+  src: url('eot?10872183');
+  src: url('eot?10872183#iefix') format('embedded-opentype'), url('woff?10872183') format('woff'), url('ttf?10872183') format('truetype'), url('svg?10872183#fontello') format('svg');
   font-weight: normal;
   font-style: normal;
 }
