+++ lib/web/static/main.js
@@ -671,7 +671,7 @@ function registerDefaultDecorators(instance) {
 
 exports.__esModule = true;
 
-var _utils = require('../utils');
+var _utils = require('../../utils');
 
 exports['default'] = function (instance) {
   instance.registerDecorator('inline', function (fn, props, container, options) {
@@ -803,7 +803,7 @@ function registerDefaultHelpers(instance) {
 
 exports.__esModule = true;
 
-var _utils = require('../utils');
+var _utils = require('../../utils');
 
 exports['default'] = function (instance) {
   instance.registerHelper('blockHelperMissing', function (context, options) {
@@ -847,7 +847,7 @@ exports.__esModule = true;
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
 
-var _utils = require('../utils');
+var _utils = require('../../utils');
 
 var _exception = require('../exception');
 
@@ -967,7 +967,7 @@ module.exports = exports['default'];
 
 exports.__esModule = true;
 
-var _utils = require('../utils');
+var _utils = require('../../utils');
 
 exports['default'] = function (instance) {
   instance.registerHelper('if', function (conditional, options) {
@@ -1040,7 +1040,7 @@ module.exports = exports['default'];
 
 exports.__esModule = true;
 
-var _utils = require('../utils');
+var _utils = require('../../utils');
 
 exports['default'] = function (instance) {
   instance.registerHelper('with', function (context, options) {
