+++ lib/web/ui/entry.hbs
