+++ lib/web/ui/js/bootstrap-modal.js
